Move packages to project owned namespace
