+++ mobile/build.gradle
@@ -5,7 +5,7 @@ android {
     compileSdkVersion 27
     buildToolsVersion '27.0.3'
     defaultConfig {
-        applicationId "com.auto.techno"
+        applicationId "io.github.markspit93.autotechno"
         minSdkVersion 21
         targetSdkVersion 27
         versionCode 1
