+++ mobile/src/main/java/io/github/markspit93/autotechno/Channel.kt
@@ -1,4 +1,4 @@
-package com.auto.techno
+package io.github.markspit93.autotechno
 
 data class Channel(
         val mediaId: String,
