+++ mobile/src/main/java/io/github/markspit93/autotechno/KotlinExt.kt
@@ -1,3 +1,3 @@
-package com.auto.techno
+package io.github.markspit93.autotechno
 
 fun <T> lazyAndroid(initializer: () -> T): Lazy<T> = lazy(LazyThreadSafetyMode.NONE, initializer)
