+++ mobile/src/main/java/io/github/markspit93/autotechno/PlayerHolder.kt
@@ -1,4 +1,4 @@
-package com.auto.techno
+package io.github.markspit93.autotechno
 
 import android.content.Context
 import android.graphics.BitmapFactory
