+++ mobile/src/test/java/io/github/markspit93/autotechno/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.auto.techno
+package io.github.markspit93.autotechno
 
 import org.junit.Test
